AbsWorker & AbsDispatcher
