+++ maro/rl_v3/training/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-from .dispatcher import Dispatcher
+from .dispatcher import TrainOpsDispatcher
 from .replay_memory import FIFOMultiReplayMemory, FIFOReplayMemory, RandomMultiReplayMemory, RandomReplayMemory
 from .train_ops import AbsTrainOps
 from .trainer import AbsTrainer, MultiTrainer, SingleTrainer, TrainerParams
@@ -9,7 +9,7 @@ from .trainer_manager import AbsTrainerManager, SimpleTrainerManager
 from .worker import TrainOpsWorker
 
 __all__ = [
-    "Dispatcher",
+    "TrainOpsDispatcher",
     "FIFOMultiReplayMemory", "FIFOReplayMemory", "RandomMultiReplayMemory", "RandomReplayMemory",
     "AbsTrainOps",
     "AbsTrainer", "MultiTrainer", "SingleTrainer", "TrainerParams",
