+++ maro/rl_v3/training/dispatcher.py
@@ -3,31 +3,12 @@
 
 from collections import deque
 
-import zmq
-from tornado.ioloop import IOLoop
-from zmq import Context
-from zmq.eventloop.zmqstream import ZMQStream
-
-from maro.rl_v3.utils.common import get_ip_address
+from maro.rl_v3.distributed import AbsDispatcher
 
 
-class Dispatcher(object):
+class TrainOpsDispatcher(AbsDispatcher):
     def __init__(self, frontend_port: int = 10000, backend_port: int = 10001) -> None:
-        # ZMQ sockets and streams
-        self._context = Context.instance()
-        self._req_socket = self._context.socket(zmq.ROUTER)
-        self._ip_address = get_ip_address()
-        self._req_socket.bind(f"tcp://{self._ip_address}:{frontend_port}")
-        self._req_endpoint = ZMQStream(self._req_socket)
-        self._dispatch_socket = self._context.socket(zmq.ROUTER)
-        self._dispatch_socket.bind(f"tcp://{self._ip_address}:{backend_port}")
-        self._dispatch_endpoint = ZMQStream(self._dispatch_socket)
-        self._event_loop = IOLoop.current()
-
-        # register handlers
-        self._dispatch_endpoint.on_recv(self._send_result_to_requester)
-
-        # workers
+        super(TrainOpsDispatcher, self).__init__(frontend_port=frontend_port, backend_port=backend_port)
         self._available_workers = deque()
         self._worker_ready = False
 
@@ -47,9 +28,3 @@ class Dispatcher(object):
         if not self._worker_ready:
             self._worker_ready = True
             self._req_endpoint.on_recv(self._route_request_to_compute_node)
-
-    def start(self) -> None:
-        self._event_loop.start()
-
-    def stop(self) -> None:
-        self._event_loop.stop()
