+++ maro/rl_v3/training/worker.py
@@ -1,23 +1,17 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-import socket
 from typing import Callable, Dict
 
-import zmq
-from tornado.ioloop import IOLoop
-from zmq import Context
-from zmq.eventloop.zmqstream import ZMQStream
-
+from maro.rl_v3.distributed import AbsWorker
 from maro.rl_v3.policy import RLPolicy
-from maro.rl_v3.utils.common import bytes_to_pyobj, bytes_to_string, pyobj_to_bytes, string_to_bytes
-from maro.utils import Logger
+from maro.rl_v3.utils.common import bytes_to_pyobj, bytes_to_string, pyobj_to_bytes
 
 from .train_ops import AbsTrainOps
 from .trainer import AbsTrainer
 
 
-class TrainOpsWorker(object):
+class TrainOpsWorker(AbsWorker):
     def __init__(
         self,
         idx: int,
@@ -26,27 +20,13 @@ class TrainOpsWorker(object):
         router_host: str,
         router_port: int = 10001
     ) -> None:
-        self._id = f"worker.{idx}"
-        self._logger = Logger(self._id)
+        super(TrainOpsWorker, self).__init__(
+            idx=idx, router_host=router_host, router_port=router_port
+        )
 
         self._policy_creator = policy_creator
         self._trainer_creator = trainer_creator
         self._trainer_dict: Dict[str, AbsTrainer] = {}
-
-        # ZMQ sockets and streams
-        self._context = Context.instance()
-        self._socket = self._context.socket(zmq.DEALER)
-        self._socket.identity = string_to_bytes(self._id)
-        self._router_ip = socket.gethostbyname(router_host)
-        self._router_address = f"tcp://{self._router_ip}:{router_port}"
-        self._socket.connect(self._router_address)
-        self._logger.info(f"Connected to dispatcher at {self._router_address}")
-        self._task_receiver = ZMQStream(self._socket)
-        self._task_receiver.send(b"READY")
-        self._event_loop = IOLoop.current()
-
-        # register handlers
-        self._task_receiver.on_recv(self._compute)
 
         self._ops_dict: Dict[str, AbsTrainOps] = {}
 
@@ -67,10 +47,4 @@ class TrainOpsWorker(object):
         self._ops_dict[ops_name].set_state(req["state"])
         func = getattr(self._ops_dict[ops_name], req["func"])
         result = func(*req["args"], **req["kwargs"])
-        self._task_receiver.send_multipart([msg[0], pyobj_to_bytes(result)])
-
-    def start(self) -> None:
-        self._event_loop.start()
-
-    def stop(self) -> None:
-        self._event_loop.stop()
+        self._receiver.send_multipart([msg[0], pyobj_to_bytes(result)])
