+++ maro/rl_v3/workflows/rollout_proxy.py
@@ -1,8 +1,8 @@
-from maro.rl_v3.rollout import BatchProxy
+from maro.rl_v3.rollout import RolloutDispatcher
 from maro.rl_v3.utils.common import from_env_as_int
 
 if __name__ == "__main__":
-    batch_proxy = BatchProxy(
+    batch_proxy = RolloutDispatcher(
         from_env_as_int("NUM_ROLLOUT_WORKERS"),
         from_env_as_int("ROLLOUT_PROXY_FRONTEND_PORT"),
         from_env_as_int("ROLLOUT_PROXY_BACKEND_PORT")
